+++ internal/pool/defaults.go
@@ -1,3 +1,11 @@
 package pool
 
-const DefaultLimit = 50
+import (
+	"time"
+)
+
+const (
+	DefaultLimit         = 50
+	defaultCreateTimeout = 5 * time.Second
+	defaultCloseTimeout  = time.Second
+)