+++ tests/slo/internal/workers/read.go
@@ -2,13 +2,12 @@ package workers
 
 import (
 	"context"
-	"fmt"
 	"math/rand"
 	"sync"
-	"time"
 
 	"golang.org/x/time/rate"
 
+	"slo/internal/log"
 	"slo/internal/metrics"
 )
 
@@ -33,7 +32,7 @@ func (w *Workers) read(ctx context.Context) (err error) {
 	defer func() {
 		m.Stop(err, attempts)
 		if err != nil {
-			fmt.Printf("[%s] get entry error: %v\n", time.Now().Format(time.RFC3339), err)
+			log.Printf("get entry error: %v", err)
 		}
 	}()